Extract tests for isArrayLike to own file
